Add files via upload
INFRASTRUCTURE AND LEARNING RESOURCES

The project focuses on enhancing the infrastructure and learning resources as outlined in NAAC Criteria 4. We aim to improve facilities such as classrooms, laboratories, seminar halls, and libraries to support effective teaching and learning. By upgrading equipment and ensuring optimal utilization of spaces, we strive to provide a conducive learning environment. Additionally, we are committed to integrating advanced technologies and resources to foster academic excellence and holistic development. To facilitate transparency and accessibility, we developed a website that displays detailed information about all the facilities and resources available. This website serves as a comprehensive platform for stakeholders to view and understand the infrastructure and learning resources of our institution.